[Clean] Rename “jsx-generator” to “code-generator”
- Rename “jsx-generator” to “code-generator"
- Move some functions to “code-generator”
